+++ src/pocketsearch/tests.py
@@ -3,7 +3,7 @@ import unittest
 import tempfile
 import datetime
 
-from pocketsearch import FileSystemReader, Text, PocketSearch, Schema, Query, Field, Int, Real, Blob, Date, Datetime, IdField
+from pocketsearch import FileSystemReader, Text, PocketSearch, Schema, Query, Field, Int, Real, Blob, Date, Datetime, IdField, Q
 
 
 class Movie(Schema):
@@ -351,43 +351,6 @@ class IndexTest(BaseTest):
         for idx, item in enumerate(self.pocket_search.search(text="is")):
             item.text  # just check, if it possible to call the attribute
 
-    def test_combine_search_results_1(self):
-        '''
-        These are all invalid queries. When we perform
-        a union order by or value arguments cannot be
-        provided. UC: order_by in second query
-        '''
-        with self.assertRaises(Query.QueryError):
-            q = self.pocket_search.search(text="paris") | self.pocket_search.search(text="england").order_by("rank")
-
-    def test_combine_search_results_2(self):
-        '''
-        These are all invalid queries. When we perform
-        a union order by or value arguments cannot be
-        provided. UC: order_by in first query
-        '''
-        with self.assertRaises(Query.QueryError):
-            q = self.pocket_search.search(text="paris").order_by("rank") | self.pocket_search.search(text="england")
-
-    def test_combine_search_results_3(self):
-        '''
-        These are all invalid queries. When we perform
-        a union order by or value arguments cannot be
-        provided.  UC: values in both queries
-        '''        
-        with self.assertRaises(Query.QueryError):
-            q = self.pocket_search.search(text="paris").values("id") | self.pocket_search.search(text="paris").values("id")
-
-    def test_combine_search_results(self):
-        # This is a correct query:
-        q1 = self.pocket_search.search(text="paris") | self.pocket_search.search(text="england")
-        self.assertEqual(q1.count(), 2)
-
-    def test_combine_search_results_using_values(self):
-        # This is a correct query:
-        q1 = self.pocket_search.search(text="paris") | self.pocket_search.search(text="england")
-        self.assertEqual(q1.values("id","text","rank").count(),2)
-
     def test_order_by_override_default_order_by(self):
         # This should override the default rank sorting by the text sorting
         r = self.pocket_search.search(text="is").values("id", "rank", "text").order_by("-text")
@@ -395,6 +358,48 @@ class IndexTest(BaseTest):
         self.assertEqual(r[1].text, self.data[2])
         self.assertEqual(r[2].text, self.data[1])
 
+class QTests(unittest.TestCase):
+    '''
+    Tests the behavior of the Q operator
+    '''
+
+    class Product(Schema):
+        '''
+        Schema with Two-field FTS index
+        '''
+        code = Text(index=True)
+        product = Text(index=True)
+        price = Int()
+
+    def setUp(self):
+        self.pocketsearch = PocketSearch(schema=self.Product)
+        for code,name,price in [
+            ("A01","Apple",4),
+            ("A02","Peach",7),
+            ("A03","Orange",8),
+            ("B01","Grapefruit",5),
+            ("B02","Banana",9),
+            ("C01","Apple and Orange",3),            
+        ]:
+            self.pocketsearch.insert(code=code,product=name,price=price)
+
+    def test_and(self):
+        q = self.pocketsearch.search(Q(product="Apple") & Q(code__allow_prefix="A*") & Q(price__gte=2))
+        self.assertEqual(q.count(),1)
+        # Test the other way around:
+        q = self.pocketsearch.search(Q(price__gte=2) & Q(product="Apple") & Q(code__allow_prefix="A*"))
+        self.assertEqual(q.count(),1)        
+        # Mixed
+        q = self.pocketsearch.search(Q(code__allow_prefix="A*") & Q(price__gte=2) & Q(product="Apple"))
+        self.assertEqual(q.count(),1) 
+
+    def test_or(self):
+        q = self.pocketsearch.search(Q(price__gte=9) | Q(price__lte=3)) 
+        self.assertEqual(q.count(),2)
+
+    def test_combined_and_or(self):
+        q = self.pocketsearch.search(Q(price__gte=1) & Q(price__lte=5) & Q(product="Apple") | Q(product="Orange"))
+        self.assertEqual(q.count(),2)
 
 class IDFieldTest(unittest.TestCase):